Move fan efficiency adjustment method to ResStockArgumentsPostHPXML.
