+++ measures/ResStockArguments/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>res_stock_arguments</name>
   <uid>c984bb9e-4ac4-4930-a399-9d23f8f6936a</uid>
-  <version_id>acb92e35-5575-4414-aefb-73efd9ff9833</version_id>
-  <version_modified>2025-11-05T17:49:07Z</version_modified>
+  <version_id>f5365c9a-4f48-44c5-a6ed-8850ed8a29f9</version_id>
+  <version_modified>2025-11-12T23:43:38Z</version_modified>
   <xml_checksum>2C38F48B</xml_checksum>
   <class_name>ResStockArguments</class_name>
   <display_name>ResStock Arguments</display_name>
@@ -11953,6 +11953,32 @@
       </choices>
     </argument>
     <argument>
+      <name>hvac_heat_pump_sizing_is_duct_limited</name>
+      <display_name>HVAC: Heat Pump Sizing Is Duct Limited</display_name>
+      <description>Whether the (ducted) heat pump has an upper limit for autosized heating/cooling capacity and an adjusted blower fan efficiency (W/CFM) value. This argument is only applicable for heat pump upgrades.</description>
+      <type>Boolean</type>
+      <required>false</required>
+      <model_dependent>false</model_dependent>
+      <choices>
+        <choice>
+          <value>true</value>
+          <display_name>true</display_name>
+        </choice>
+        <choice>
+          <value>false</value>
+          <display_name>false</display_name>
+        </choice>
+      </choices>
+    </argument>
+    <argument>
+      <name>baseline_max_airflow_cfm</name>
+      <display_name>baseline_max_airflow_cfm</display_name>
+      <description>TODO.</description>
+      <type>Double</type>
+      <required>false</required>
+      <model_dependent>false</model_dependent>
+    </argument>
+    <argument>
       <name>hvac_blower_fan_watts_per_cfm</name>
       <display_name>HVAC: Blower Fan Efficiency</display_name>
       <description>The blower fan efficiency at maximum fan speed. Applies only to split (not packaged) systems (i.e., applies to ducted systems as well as ductless mini-split systems)..</description>
@@ -12926,7 +12952,7 @@
       <filename>README.md</filename>
       <filetype>md</filetype>
       <usage_type>readme</usage_type>
-      <checksum>142844D6</checksum>
+      <checksum>35E3995D</checksum>
     </file>
     <file>
       <filename>README.md.erb</filename>
@@ -12943,7 +12969,7 @@
       <filename>measure.rb</filename>
       <filetype>rb</filetype>
       <usage_type>script</usage_type>
-      <checksum>FD56B63D</checksum>
+      <checksum>E0F7BA47</checksum>
     </file>
     <file>
       <filename>constants.rb</filename>
